--- WebCrawlers/FileTree.md
@@ -1,9 +1,0 @@
-.
-├── a.out  //可执行文件
-├── FileTree.md//文件树，文件夹内的文件进行简短的用途介绍
-├── mysqlplus.c//对mysql操作的一些函数实现,包括上传，删除，查找，下载等
-├── mysqlplus.h//库函数，包括一些关于mysql操作的结构体和函数的定义
-├── readme.md//关于爬虫思路的简单介绍
-└── regexbook4.c//main函数，4表明迭代开发的第四次迭代后的产物.
-
-0 directories, 6 files
